requirements.txt
textblob
re
discord.ext
json
datetime
ollama
random
difflib
dotenv
discord.py
asyncio
pathlib
subprocess
os
spacy
